Удалил TailPart.
Ошибки компиляции:
Severity	Code	Description
Error		Could not copy the file "\snake_game\Content\bin\DejaVu Sans Mono.xnb" because it was not found.
Error		Could not copy the file "\snake_game\Content\bin\images\head.xnb" because it was not found.
Error		Could not copy the file "\snake_game\Content\bin\images\apple.xnb" because it was not found.
Error		Could not copy the file "\snake_game\Content\bin\images\snake.xnb" because it was not found.
Error		Could not copy the file "\snake_game\Content\bin\images\brick.xnb" because it was not found.